Sets timer back to all zeroes when countdown is stopped
Previously, when you'd force touch on the app to pop the contextual menu and then choose Stop, it would simply arrest the countdown timer at the current value.  Now it resets the countdown timer to zero.
Also, it turns out this all *was* working in the prior commit.  w00t!